refactor: relocate `com.github.learndifferent.mtm.chain` to `com.github.learndifferent.mtm.chain.scraper`
